Bugfix on workoutdetails

+++ app/src/main/java/com/example/justrun/activities/WorkoutDetailsActivity.kt
@@ -50,7 +50,7 @@ class WorkoutDetailsActivity : AppCompatActivity() {
     private fun formatDistance(workout: WorkoutData): String {
         val distance = workout.distance
         return if (distance <= 100) {
-            getString(R.string.distance_m, distance)
+            getString(R.string.distance_m, distance.toInt())
         } else {
             val distanceKilometers = distance.toDouble().div(1000)
             getString(R.string.distance_km, distanceKilometers)
